requirements.txt
Django==2.2.7
django-leaflet~=0.27.1
pandas~=1.1.3
plotly~=4.11.0
djangorestframework~=3.12.1
django-geojson~=3.2.1
django-bootstrap4~=22.1
django-plotly-dash~=1.7.2
dpd-static-support~=0.0.5
mysqlclient~=2.1.0
dash~=1.20.0
dash-bootstrap-components==0.13.1
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
dpd-components~=0.1.0
numpy~=1.22.4
sqlparse~=0.4.2
whitenoise~=6.1.0
django-colorfield~=0.7.0
